+++ CHANGES.md
@@ -1,9 +1,11 @@
+
 ChangeLog
 =========
 
 Development 0.1.0
 ------------------
 
+- Enabled suffix text labels for input fields [#71,#74]
 - Add support for custom data attributes on file cards [#73]
 - Truncate long file names and titles on file cards [#70]
 - Render preview of file cards in a modal dialog [#65]